codevoice-backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
openai-whisper==20231117
torch>=1.10
torchaudio>=0.10
transformers==4.41.2
spacy==3.7.4
ollama==0.1.7
requests
numpy
soundfile
webrtcvad==2.0.10
